refactor(proxy): extract DML handlers
Move write_row, update_row, and delete_row handling into dml.cc.
Register dml.cc in the proxy build.
