+++ proxy/ha_lineairdb.cc
@@ -1188,177 +1188,6 @@ int ha_lineairdb::index_read_last(uchar *buf, const uchar *key, uint key_len) {
   return index_read_map(buf, key, keypart_map, HA_READ_PREFIX_LAST);
 }
 
-/**
-  @brief
-  write_row() inserts a row.
-  No extra() hint is given currently if a bulk load is happening.
-  @param buf is a byte array of data.
-*/
-int ha_lineairdb::write_row(uchar *buf) {
-  DBUG_TRACE;
-
-  set_write_buffer(buf);
-  auto key = extract_key(buf);
-
-  auto tx = get_transaction(ha_thd());
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  // buffer_write appends to a local buffer (no RPC yet), so no error check needed.
-  // The actual RPC is sent at flush time (buffer full, table change, or commit).
-  tx->buffer_write(db_table_name, key, write_buffer_);
-
-  // Write secondary index entries.
-  // Normal transactions check UNIQUE indexes immediately.
-  // Prefetch sends UNIQUE index writes to validate-and-commit with row writes.
-  for (uint i = 0; i < table->s->keys; i++) {
-    auto key_info = table->key_info[i];
-    if (i == table->s->primary_key) continue;
-
-    std::string secondary_key = build_secondary_key_from_row(buf, key_info);
-
-    if (key_info.flags & HA_NOSAME) {
-      if (tx->is_prefetch_mode()) {
-        tx->buffer_write_secondary_index(db_table_name, key_info.name,
-                                         secondary_key, key);
-      } else {
-        tx->flush_write_buffer();
-        tx->choose_table(db_table_name);
-        bool ok = tx->write_secondary_index(key_info.name, secondary_key, key);
-        if (!ok || tx->is_aborted()) {
-          return abort_errno(tx);
-        }
-      }
-    } else {
-      tx->buffer_write_secondary_index(db_table_name, key_info.name,
-                                        secondary_key, key);
-    }
-  }
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  tx->add_rowcount_delta(share, db_table_name, +1);
-
-  return 0;
-}
-
-int ha_lineairdb::update_row(const uchar *old_data, uchar *new_data) {
-  DBUG_TRACE;
-
-  auto tx = get_transaction(ha_thd());
-  auto key = extract_key_from_mysql(old_data);
-  const auto new_key = extract_key_from_mysql(new_data);
-
-  // FIXME: reject a PK-changing UPDATE. update_row overwrites in place at the old
-  // key and cannot move a row, so executing one would store new_data under the
-  // old key with nothing at the new key -- silent corruption. A real move (delete
-  // old + insert new + secondary-index rewrite) is not implemented.
-  if (key != new_key) {
-    return prefetch_reject_unsupported(ha_thd(), tx,
-                                       "primary-key-changing UPDATE");
-  }
-
-  if (key.empty()) {
-    key = last_fetched_primary_key_;
-  }
-
-  if (key.empty()) {
-    key = extract_primary_key_from_ref(ref);
-  }
-
-  last_fetched_primary_key_ = key;
-
-  set_write_buffer(new_data);
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  // Buffer the base-row update; read/scan paths and commit flush it later.
-  tx->buffer_write(db_table_name, key, write_buffer_);
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  for (uint i = 0; i < table->s->keys; i++) {
-    auto key_info = table->key_info[i];
-
-    if (i == table->s->primary_key) {
-      continue;
-    }
-
-    std::string old_secondary_key =
-        build_secondary_key_from_row(old_data, key_info);
-    std::string new_secondary_key =
-        build_secondary_key_from_row(new_data, key_info);
-
-    if (old_secondary_key == new_secondary_key) {
-      continue;
-    }
-
-    tx->update_secondary_index(key_info.name, old_secondary_key,
-                               new_secondary_key, key);
-
-    if (tx->is_aborted()) {
-      return abort_errno(tx);
-    }
-  }
-
-  return 0;
-}
-
-int ha_lineairdb::delete_row(const uchar *buf) {
-  DBUG_TRACE;
-
-  auto key = extract_key_from_mysql(buf);
-
-  if (key.empty()) {
-    key = last_fetched_primary_key_;
-  }
-
-  if (key.empty()) {
-    return HA_ERR_KEY_NOT_FOUND;
-  }
-
-  last_fetched_primary_key_ = key;
-
-  auto tx = get_transaction(ha_thd());
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  // Buffer the base-row delete; read/scan paths and commit flush it later.
-  tx->buffer_delete(db_table_name, key);
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  for (uint i = 0; i < table->s->keys; i++) {
-    auto key_info = table->key_info[i];
-    if (i != table->s->primary_key) {
-      std::string secondary_key = build_secondary_key_from_row(buf, key_info);
-
-      tx->buffer_delete_secondary_index(db_table_name, key_info.name,
-                                        secondary_key, key);
-    }
-  }
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  tx->add_rowcount_delta(share, db_table_name, -1);
-
-  return 0;
-}
-
 int ha_lineairdb::index_read_map(uchar *buf, const uchar *key,
                                  key_part_map keypart_map,
                                  enum ha_rkey_function find_flag) {
